integrate icns name service to groups page
